requirements.txt
streamlit
pandas
numpy
plotly
scipy
fpdf2